Improve AI by considering all words, sometimes
Have AI consider all word choices when the valid word set is small. If
time were not a consideration, I suppose I might have it consider all
words for every guess. But also, this mode switching is how I'm starting
to play as a human, so it's nice for it to line up. (Although, I
probably only switch to all words when the word count is <=10 instead of
<=30)