Better performance on quordle+: prioritize known
No matter what, once a word is narrowed down, we have to guess it at
least once. So by then, we might as well guess it early, in order to see
what clues it gives us about the remaining words.

Plus a bonus of showing the number of answers, to help avoid mistakes
when entering in octordle at CLI.